services/driver_location/requirements.txt
fastapi
uvicorn
redis[async]
pydantic<2
python-jose[cryptography]
python-dotenv
